requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
decorator==5.1.1
Flask==2.0.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.6
MarkupSafe==2.0.1
numba==0.54.1
numpy==1.18.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pooch==1.5.2
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
